# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/devices": {
            "get": {
                "description": "Retrieves a high-performance chunk of devices starting after a target cursor ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Fetch all devices with cursor pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items to return (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID of the last item from the previous page (default 0)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CursorPageResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            },
            "post": {
                "description": "Persists a device into the database with an automatic creation time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Create a new device",
                "parameters": [
                    {
                        "description": "Device Input Payload Data",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeviceDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            }
        },
        "/devices/search/brand": {
            "get": {
                "description": "Filters the device collection case-insensitively by brand matching text.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Fetch devices by brand with pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the device brand",
                        "name": "brand",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Device"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            }
        },
        "/devices/search/state": {
            "get": {
                "description": "Filters the device collection matching the target state enum value.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Fetch devices by state with pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target state (AVAILABLE, IN_USE, INACTIVE)",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Device"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "Retrieves a specific device resource matching the unique path variable identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Fetch a single device by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the target device to find",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            },
            "put": {
                "description": "Replaces all fields of an existing device record. Protected by active state guards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Fully update an existing device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the target device",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Complete new state data payload",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeviceDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently removes a specific device matching the path identifier key. Protected by active state guards.",
                "tags": [
                    "Devices API"
                ],
                "summary": "Delete a single device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the target device to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "24": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Modifies specific payload key-value fields. Protected by active state guards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices API"
                ],
                "summary": "Partially update an existing device",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the target device",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Key-value properties map to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Device"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ApiError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.CursorPageResult": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Device"
                    }
                },
                "hasNext": {
                    "type": "boolean"
                },
                "nextCursor": {
                    "type": "integer"
                }
            }
        },
        "model.Device": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "creationTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/model.DeviceState"
                }
            }
        },
        "model.DeviceDTO": {
            "type": "object",
            "required": [
                "brand",
                "name",
                "state"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/model.DeviceState"
                }
            }
        },
        "model.DeviceState": {
            "type": "string",
            "enum": [
                "AVAILABLE",
                "IN_USE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "Available",
                "InUse",
                "Inactive"
            ]
        }
    }
}